+++ CHANGELOG.md
@@ -22,3 +22,4 @@ All changes
   - [feat] add erase to end of line api
   - [feat] add EOF_ to EOF;SIGINT_ to SIGINT auto convert
   - [fix] chry_readline arg3 arg4 change uint32_t to uint16_t
+  - [feat] add auto_refresh control